Use new identifier forest API
